more changes; refactors chart functions into the chart reducer
