+++ src/helpers/isColTypeTest.js
@@ -1,7 +1,11 @@
 
 module.exports = function isColTypeTest (selectedColumnDef, dataTypeToTest) {
-  if (selectedColumnDef.type === dataTypeToTest) {
-    return true
+  if(typeof selectedColumnDef !== 'undefined' && selectedColumnDef){
+    if(Object.keys(selectedColumnDef).indexOf('type') > -1){
+      if (selectedColumnDef.type === dataTypeToTest) {
+        return true
+      }
+    }
   }
   return false
 }
