+++ src/helpers/roundNumberByPower.js
@@ -1,0 +1,3 @@
+module.exports = function roundNumberByPower (num, tensBaseToRoundTo) {
+    return Math.ceil(num/tensBaseToRoundTo)*tensBaseToRoundTo
+}
